+++ editor/editor-plugin-block-controls/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/editor-plugin-block-controls
 
+## 7.7.7
+
+### Patch Changes
+
+- [`8854ad2383b33`](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/commits/8854ad2383b33) -
+  Suppress no-literal-string-in-jsx
+- Updated dependencies
+
 ## 7.7.6
 
 ### Patch Changes