+++ editor/editor-plugin-highlight/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atlaskit/editor-plugin-highlight
 
+## 6.3.9
+
+### Patch Changes
+
+- [`8854ad2383b33`](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/commits/8854ad2383b33) -
+  Suppress no-literal-string-in-jsx
+
 ## 6.3.8
 
 ### Patch Changes