+++ editor/editor-plugin-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugin-table
 
+## 15.4.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 15.4.2
 
 ### Patch Changes